Mouse Rotation
Two Scripts, one defines user targeting in world space with a mouse (attach that to a camera). The other script rotates any game object to point to where the mouse points to in world space (attach to mitch, etc).

Also renames scripts to follow naming convention, may need to re-add scripts to objects if errors appear.